Move headers to regression test
